refactor(dbt-export): extract field→test mapping helper
Pulls _to_column's per-field test logic out of dbt_exporter.py into a new
datacontract/integration/dbt_test_mapping.py module that exposes a single
field_to_data_tests() function. No behavior change — existing exporter
golden-file tests pass unchanged. Preparatory for a future dbt sync
command that will reuse the same mapping.
